warm: the index prices the buildings, and calibration stops costing a walk
A 451-node import took 35 minutes to warm, and the operator said so. The
profile of a real warm answered with one line: 61% of the whole run inside
environ's corridor scan, and most of THAT inside the runtime zeroing freshly
allocated memory - a fresh seen-map and a bounding-box tile scan per corridor
step, per pair, 101,010 times. The indexed twin - bucketed footprints, an
epoch-marked scratch, nothing allocated on the hot path - existed the whole
time, held to the same crossings by TestPathIndexMatchesDirect; the coverage
rasters were already using it. Only the two paths that price every pair were
not: the engine's own buildingLossDB and the GPU warm's pricing pass. Both go
through the index now.

And a calibration no longer throws the matrix away. The excess-loss term was
baked into every cached entry and hashed into the matrix fingerprint, so
validate.calibrate - a constant every path shares - invalidated half an hour
of ground-walking to move one number. The cache now stores the raw physics
and the term is applied where the cache is read: two engines sharing one
matrix but carrying different terms answer exactly the term apart, and
TestExcessLossIsReadNotStored holds them to it without a single profile
walked. Old matrices, keyed under fingerprints that included the term, never
match the new fingerprint and are remeasured once. It also ends a quiet
inconsistency: culled below-floor entries never carried the term while
measured ones did, and now neither does.

The default the term starts from is the world calibration itself:
DefaultExcessLossDB moves from 20 to 23.5, the total fitted against 1,330
receptions across the whole of ScotMesh rather than 118 across Fife. It is
what a network nobody has observed gets - a clutter term for UK terrain at
869 MHz, not a ScotMesh setting - and validate.calibrate refines it wherever
observations exist.

Also swept: extraction directories the environment downloader leaves behind
when the process dies mid-pull. The cleanup closure handles every ordinary
exit, a killed process cannot run it, and on Linux the temp directory is
tmpfs - three abandoned extractions were half a gigabyte of somebody's RAM.

The terrain benches that bracketed the diagnosis stay: they measure the
profile walk against a real tile store, serial and shared, and skip without
one. They are what said the store was innocent - 205,000 samples a second
across twelve workers - and pointed the profiler at the corridor instead.

+++ internal/app/session/validate.go
@@ -245,7 +245,7 @@ func registerValidate(st *state.Store, s *Sim) {
 			w.Links = nil
 			s.warm(st, len(s.nodes))
 		}
-		w.Say(fmt.Sprintf("excess path loss back to the %d dB default", DefaultExcessLossDB))
+		w.Say(fmt.Sprintf("excess path loss back to the %.1f dB default", float64(DefaultExcessLossDB)))
 		return map[string]any{"db": DefaultExcessLossDB}, nil
 	})
 }
